% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{displayNamesDF}
\alias{displayNamesDF}
\title{Variable Display Names}
\format{A data frame with 4 columns}
\usage{
displayNamesDF
}
\description{
A dataset containing the full names and units of parameters used in the model
}
\keyword{datasets}
