% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FiltNumericColumns}
\alias{FiltNumericColumns}
\title{Produces summary statistics for an entire numeric dataframe}
\usage{
FiltNumericColumns(dataframe)
}
\arguments{
\item{dataframe}{the dataframe}
}
\value{
a dataframe of summary statistics
}
\description{
Produces summary statistics for an entire numeric dataframe
}
