% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FindBestModel}
\alias{FindBestModel}
\title{Runs a stepwise optimisation to select the best fitting parameters}
\usage{
FindBestModel(df, variableList, direction = "backward", steps = 1000)
}
\arguments{
\item{df}{a dataframe}

\item{variableList}{a list of parameters from the dataframe}

\item{direction}{"forward" or "backwards. Default "backward"}

\item{steps}{the number of steps for the model to assess}
}
\description{
Runs a stepwise optimisation to select the best fitting parameters
}
