% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ksource}
\alias{ksource}
\title{Runs an external Knitr script}
\usage{
ksource(x, ...)
}
\arguments{
\item{x}{A filepath to an external knitr script}

\item{...}{options to be passed to the purl function}
}
\description{
Runs an external Knitr script
}
