% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticDiagnostics}
\alias{LogisticDiagnostics}
\title{Logistic Regression Model Diagnostics}
\usage{
LogisticDiagnostics(Model)
}
\arguments{
\item{Model}{A glm object}
}
\value{
A printout of diagnostics
}
\description{
Prints results for the Chi Squared, Psuedo R sqaured values, Variance Inflation
Factors and Durbin Watson Test
}
