% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticModelInt}
\alias{LogisticModelInt}
\title{Takes a list of variables, and  updates the list to include the log transformations (Int)}
\usage{
LogisticModelInt(PredictorVariables, outcome, df)
}
\arguments{
\item{PredictorVariables}{a list of varaibles to be assessed}

\item{outcome}{the outcome variable}

\item{df}{a datafram containing the observations}
}
\description{
Takes a list of variables, and  updates the list to include the log transformations (Int)
}
