% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticOddsTable}
\alias{LogisticOddsTable}
\title{Builds a Logistic Odds Table}
\usage{
LogisticOddsTable(Model, Sort = FALSE)
}
\arguments{
\item{Model}{a glmmodel}

\item{Sort}{boolean indicator whether the parameters should be sorted. Default FALSE}
}
\description{
Produces an odds table with confidence intervals for a logistic model
}
