% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticPseudoR2s}
\alias{LogisticPseudoR2s}
\title{Calculate Logistic Model R-squared values}
\usage{
LogisticPseudoR2s(Model)
}
\arguments{
\item{Model}{A glm object}
}
\description{
Calculates the R squared values (Hosmer and Lemeshow, Cox and Snell and Nagelkerke) values for a logistic regression model.
}
