% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogisticResultsTable}
\alias{LogisticResultsTable}
\title{Creates a formatted odds table of the results using ggplot}
\usage{
LogisticResultsTable(Model, roundby = 3)
}
\arguments{
\item{Model}{a logistic regression object}

\item{roundby}{the number of decimal places to display results. Default 3}
}
\description{
Creates a formatted odds table of the results using ggplot
}
