% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogOddsPlotGraph}
\alias{LogOddsPlotGraph}
\title{Produces boxplots for estimated values from a regression model}
\usage{
LogOddsPlotGraph(OddsTable, PlotTitle, Sort = FALSE)
}
\arguments{
\item{OddsTable}{a formatted odds table from the function "LogisticOddsTable"}

\item{PlotTitle}{the title of the resulting plot}

\item{Sort}{reorder the plot by variable fit}
}
\description{
Produces boxplots for estimated values from a regression model
}
