% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LogResults}
\alias{LogResults}
\title{Summarise results for a logistic odds model}
\usage{
LogResults(Model)
}
\arguments{
\item{Model}{a glm object}
}
\description{
Summarise results for a logistic odds model
}
