% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matchNamesColumns}
\alias{matchNamesColumns}
\title{Checks the row names against the variable full names}
\usage{
matchNamesColumns(dataframe, displayNamesDF, dropNA = TRUE)
}
\arguments{
\item{dataframe}{a list of names which are to looked up}

\item{displayNamesDF}{the path of the reference file}

\item{dropNA}{a TRUE/FALSE selection whether names should be dropped if empty}
}
\description{
This function is designed to work with the PhD project
}
