% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ModelAccuracyFromModel}
\alias{ModelAccuracyFromModel}
\title{Assess the model accuracy}
\usage{
ModelAccuracyFromModel(Model, Dataframe, ...)
}
\arguments{
\item{Model}{the glm object}

\item{Dataframe}{the full dataframe used to build the model}

\item{...}{additional arguments passed to the `ModelAccuracy` function.}
}
\description{
Wrapper for the ModelAccuracy function to extract the variables from the model
  Would probably be better to rewrite the original function
}
