% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ParameterUpdate}
\alias{ParameterUpdate}
\title{Update the list of parameters for the model building. Used in conjunction with the
"LogisticModel" and "LogsticModelInt" functions}
\usage{
ParameterUpdate(input, add = NULL, remove = NULL)
}
\arguments{
\item{input}{an existing list of input parameters}

\item{add}{a list of parameters to be added to the model (optional)}

\item{remove}{a list of values to be removed from the model (optional)}
}
\value{
An updateded model parameter list
}
\description{
Update the list of parameters for the model building. Used in conjunction with the
"LogisticModel" and "LogsticModelInt" functions
}
