% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ReducedDataframe}
\alias{ReducedDataframe}
\title{Reduces a dataframe to a list of parameters for use within regression modelling}
\usage{
ReducedDataframe(dataframe, outcomeVariable, inputVariables)
}
\arguments{
\item{dataframe}{a dataframe}

\item{outcomeVariable}{the outcome variable}

\item{inputVariables}{a list of input variables}
}
\description{
Reduces a dataframe to a list of parameters for use within regression modelling
}
