% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Run Analysis}
\usage{
runAnalysis(cache = "_code/outputs/codeOutputs.RData",
  analysis = "AnalysisComplete.Rmd")
}
\arguments{
\item{cache}{the filepath of the cached file}

\item{analysis}{the filepath of the analysis file}
}
\description{
Runs the full statistical analysis
}
