% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{A fill scale fill for the publication}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{option arguments passed to `discrete_scale`}
}
\description{
A fill scale fill for the publication
}
