% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ScatterPlotOdds}
\alias{ScatterPlotOdds}
\title{Produces a scatter plot for the logistic regression variables}
\usage{
ScatterPlotOdds(df, variable, quantiles = 20)
}
\arguments{
\item{df}{the dataframe which contains the data}

\item{variable}{the column reference to search}

\item{quantiles}{the number of bins in which the data should be grouped}
}
\description{
Produces a scatter plot for the logistic regression variables
}
