% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_Dataset}
\alias{Segmented_Dataset}
\title{Splits a dataset based on a categorical variable}
\usage{
Segmented_Dataset(df, by)
}
\arguments{
\item{df}{a dataframe}

\item{by}{the column used to split the data}
}
\value{
a list of dataframes
}
\description{
Splits a dataset based on a categorical variable
}
