% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_LogisticModels}
\alias{Segmented_LogisticModels}
\title{Build Segmented Logistic Regression Models}
\usage{
Segmented_LogisticModels(SegmentedDatasets, predictors, outcome)
}
\arguments{
\item{SegmentedDatasets}{the full dataframe to be split}

\item{predictors}{the name of the column for the dataset to be split using}

\item{outcome}{the outcome variable}
}
\value{
a list of logistic regression models
}
\description{
Combines the SplitDatasetbyVariable function and "Logstic Model"
}
