% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_LogisticModelsComplete}
\alias{Segmented_LogisticModelsComplete}
\title{Returns the a list of results for split logistic regresion model.}
\usage{
Segmented_LogisticModelsComplete(df, split_category, variables_list,
  outcome_variable, linebreak = 0.1, limits = 1)
}
\arguments{
\item{df}{a dataframe to be split}

\item{split_category}{the column to be used to split the dataset by}

\item{variables_list}{a list of variables to be used within the regression model}

\item{outcome_variable}{the outcome variable}

\item{linebreak}{the spacing to be used between gridlines in the plot}

\item{limits}{the extent of the plot}
}
\description{
This function splits data based on a categorical variable, builds separate
  glm models for each dataset and compares the results using a Odds Plot.
  This returns the full stages of the analysis 1) dataset 2) glm models
  3) Summary Odds table 4) logistic regresion plot.
}
