% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_OddsPlot}
\alias{Segmented_OddsPlot}
\title{Builds an odds plot for a segmented logistic regression model}
\usage{
Segmented_OddsPlot(OddsTables, PlotTitle, linebreak,
  displayNamesDF = WindAnalysis::displayNamesDF)
}
\arguments{
\item{OddsTables}{a list of odds as produced from the function "OddsTableSegmented"}

\item{PlotTitle}{the title for the ggplot object}

\item{linebreak}{numeric. Spacing between minor line breaks in plot}
}
\description{
Builds an odds plot for a segmented logistic regression model
}
