% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Segmented_OddsPlotGroupedCustom}
\alias{Segmented_OddsPlotGroupedCustom}
\title{Plots a faceted odds ratio plot for a list of segmented odds tables}
\usage{
Segmented_OddsPlotGroupedCustom(OddsTables, linebreak = 0.2, scale = 1)
}
\arguments{
\item{OddsTables}{a list of odds as produced from the function "OddsTableSegmented"}

\item{linebreak}{numeric. Spacing between minor line breaks in plot}

\item{scale}{the limits of the plot}
}
\description{
Plots a faceted odds ratio plot for a list of segmented odds tables
}
