% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_Publication}
\alias{theme_Publication}
\title{A theme for use within the publication}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{the font size of the plot Defaults to 14}

\item{base_family}{the font family of the plot. Defaults to "helvetica"}
}
\description{
A theme for use within the publication
}
