% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TwoWayFrequency}
\alias{TwoWayFrequency}
\title{Creates a 2 way frequency table from a dataframe and calculates the percentage of
a specified category}
\usage{
TwoWayFrequency(Rows, Columns, SumPercentage, inputDataframe)
}
\arguments{
\item{Rows}{the variable to counted in the rows}

\item{Columns}{the variable to be counted in the columns}

\item{SumPercentage}{the parameter to be summed in the column}

\item{inputDataframe}{the dataframe containing the row and column parameters}
}
\description{
Creates a 2 way frequency table from a dataframe and calculates the percentage of
a specified category
}
