% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VariableCluster}
\alias{VariableCluster}
\title{Divides a set of numeric variables into disjoint or hierarchical cluster
  which can be used to diagnose collinearity between variables}
\usage{
VariableCluster(PredictorVariables, df)
}
\arguments{
\item{PredictorVariables}{Parameter names to be considered}

\item{df}{The dataframe containing the parameters}
}
\description{
Divides a set of numeric variables into disjoint or hierarchical cluster
  which can be used to diagnose collinearity between variables
}
