% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VIFcheck}
\alias{VIFcheck}
\title{Calculate Variance Inflation Factor (VIF)}
\usage{
VIFcheck(Model)
}
\arguments{
\item{Model}{an object that responds to coef, vcov, and model.matrix, such as an lm or glm object}
}
\description{
Calculates variance-inflation and generalized variance-inflation factors for
   linear and generalized linear models, and returns any which are above 10.
}
